test: tidy in tests
