+++ src/postgres/queries/mod.rs
@@ -4,3 +4,5 @@ pub mod get_tables;
 mod test_get_enums;
 #[cfg(test)]
 mod test_get_tables;
+#[cfg(test)]
+mod test_helper;
